+++ src/loaders/types.js
@@ -1,3 +1,5 @@
+import { FileType } from '../app/constants';
+
 import JsonLoader from './JsonLoader';
 import MatrixZarrLoader from './MatrixZarrLoader';
 import GenesJsonAsMatrixZarrLoader from './GenesJsonAsMatrixZarrLoader';
@@ -12,24 +14,21 @@ import CellsGeoJsonLoader from './CellsGeoJsonLoader';
 
 import { AnnDataSource, ZarrDataSource, JsonSource } from './data-sources';
 
-const ANNDATA = 'anndata';
-
-const fileTypeToLoaderAndSource = {
-  'expression-matrix.zarr': [ZarrDataSource, MatrixZarrLoader],
-  'clusters.json': [JsonSource, ClustersJsonAsMatrixZarrLoader],
-  'genes.json': [JsonSource, GenesJsonAsMatrixZarrLoader],
-  'cells.json': [JsonSource, JsonLoader],
-  'molecules.json': [JsonSource, MoleculesJsonLoader],
-  'neighborhoods.json': [JsonSource, JsonLoader],
-  'raster.json': [JsonSource, RasterJsonLoader],
-  'raster.ome-zarr': [ZarrDataSource, OmeZarrLoader],
-  'cell-sets.json': [JsonSource, CellSetsJsonLoader],
-  [`${ANNDATA}-cell-sets.zarr`]: [AnnDataSource, AnnDataLoaders.CellSetsZarrLoader],
-  [`${ANNDATA}-cells.zarr`]: [AnnDataSource, AnnDataLoaders.CellsZarrLoader],
-  [`${ANNDATA}-expression-matrix.zarr`]: [AnnDataSource, AnnDataLoaders.MatrixZarrLoader],
-  [`${ANNDATA}-molecules.zarr`]: [AnnDataSource, AnnDataLoaders.MoleculesZarrLoader],
-  'genomic-profiles.zarr': [ZarrDataSource, GenomicProfilesZarrLoader],
-  'cells.geojson': [JsonSource, CellsGeoJsonLoader],
+export const fileTypeToLoaderAndSource = {
+  [FileType.EXPRESSION_MATRIX_ZARR]: [ZarrDataSource, MatrixZarrLoader],
+  [FileType.CLUSTERS_JSON]: [JsonSource, ClustersJsonAsMatrixZarrLoader],
+  [FileType.GENES_JSON]: [JsonSource, GenesJsonAsMatrixZarrLoader],
+  [FileType.CELLS_JSON]: [JsonSource, JsonLoader],
+  [FileType.MOLECULES_JSON]: [JsonSource, MoleculesJsonLoader],
+  [FileType.NEIGHBORHOODS_JSON]: [JsonSource, JsonLoader],
+  [FileType.RASTER_JSON]: [JsonSource, RasterJsonLoader],
+  [FileType.RASTER_OME_ZARR]: [ZarrDataSource, OmeZarrLoader],
+  [FileType.CELL_SETS_JSON]: [JsonSource, CellSetsJsonLoader],
+  [FileType.ANNDATA_CELL_SETS_ZARR]: [AnnDataSource, AnnDataLoaders.CellSetsZarrLoader],
+  [FileType.ANNDATA_CELLS_ZARR]: [AnnDataSource, AnnDataLoaders.CellsZarrLoader],
+  [FileType.ANNDATA_EXPRESSION_MATRIX_ZARR]: [AnnDataSource, AnnDataLoaders.MatrixZarrLoader],
+  [FileType.GENOMIC_PROFILES_ZARR]: [ZarrDataSource, GenomicProfilesZarrLoader],
+  [FileType.CELLS_GEOJSON]: [JsonSource, CellsGeoJsonLoader],
 };
 
 export function getSourceAndLoaderFromFileType(type) {